requirements.txt
alembic==1.2.0
APScheduler==3.6.1
astroid==1.6.6
autopep8==1.4.4
backports.functools-lru-cache==1.5
blinker==1.4
boto3==1.9.235
botocore==1.12.235
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
configparser==3.8.1
docutils==0.15.2
enum34==1.1.6
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.4.0
funcsigs==1.0.2
futures==3.3.0
google-api-core==1.3.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-storage==1.11.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
GoogleAppEngineCloudStorageClient==1.9.22.1
gunicorn==19.9.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lazy-object-proxy==1.4.2
lib==3.0.0
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
protobuf==3.9.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pylint==1.9.5
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.8
typing==3.7.4.1
tzlocal==2.0.0
urllib3==1.25.6
Werkzeug==0.15.6
wrapt==1.11.2
